nycrashes/backend/chat/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.118.0",
    "strands-agents>=1.10.0",
    "uvicorn>=0.37.0",
]
